Deleting old websites + Using localStorage
Deleting V1 and V2 folders. Thinking about it, they're probably just in old revisions of this repo. LOL.

Also restructuring the main.js file and taking a huge chunk of it and storing it in commands.js. "commands.js" is a js file that has helpers and all the commands (along with dictionaries of them).

Also revamped the commands system. Instead of hard coding them, the simple link redirects are no longer default. Instead, the user can create them through a command, which they will be saved and read from local storage. Allows for nice flexibility with the code, not having to code everytime I want to add a new command.